完善对账的查询以及添加补偿功能

+++ damai-server/damai-order-service/src/main/java/com/damai/domain/SeatRecord.java
@@ -13,6 +13,6 @@ public class SeatRecord {
     private Long ticketCategoryId;
     private Long seatId;
     private Long ticketUserId;
-    private Long beforeStatus;
-    private Long afterStatus;
+    private Integer beforeStatus;
+    private Integer afterStatus;
 }
